Service Commit - Renamed utils to helper_functions to avoid confusing it with utility.py - Made ping-command an ephemeral message
