+++ src/commands/dm_tools.py
@@ -29,7 +29,7 @@ from discord.commands import (
 )
 import discord
 from discord.ext import commands
-from utils import get_project_root, from_project_root, config_var
+from helper_functions import get_project_root, from_project_root, config_var
 
 roles_path = os.path.join(get_project_root(), "/config/roles.json")
 with open(from_project_root('/config/roles.json'), encoding='utf-8') as roles_json:
